+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## [0.10.1] - 2026-07-16
+
+### Fixed
+
+- Webfinger, ActivityPub clients timeout after 30 seconds
+- Better error reporting in crawler worker
+
 ## [0.10.0] - 2026-07-16
 
 ### Changed
@@ -262,7 +271,8 @@ keepalive connections in ActivityPubClient and WebfingerClient
 - Crawler max depth defaults to 1.
 - `DatabaseGraph` no longer uses its own transaction.
 
-[Unreleased]: https://github.com/pub-crawler/pub-crawler/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/pub-crawler/pub-crawler/compare/v0.10.1...HEAD
+[0.10.1]: https://github.com/pub-crawler/pub-crawler/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/pub-crawler/pub-crawler/compare/v0.9.2...v0.10.0
 [0.9.2]: https://github.com/pub-crawler/pub-crawler/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/pub-crawler/pub-crawler/compare/v0.9.0...v0.9.1